create page controller and place all main render logic inside that controller
